+++ rmf_traffic/src/rmf_traffic/reservations/internal/Protocol.hpp
